move job offers to compensation
