+++ packages/core/src/api.ts
@@ -10,10 +10,10 @@ dayjs.extend(timezone);
 import { airtableWorker } from './modules/airtable/airtable.core';
 import { applicationWorker } from './modules/application/application.core';
 import { oneTimeCodeWorker } from './modules/authentication/one-time-code.worker';
+import { jobOfferWorker } from './modules/compensation/job-offers';
 import { eventWorker } from './modules/event/event.worker';
 import { feedWorker } from './modules/feed/feed';
 import { gamificationWorker } from './modules/gamification/gamification.core';
-import { jobOfferWorker } from './modules/job-offers/job-offers';
 import { mailchimpWorker } from './modules/mailchimp/mailchimp';
 import { memberEmailWorker } from './modules/member/member-email.worker';
 import { memberWorker } from './modules/member/member.worker';
