+++ packages/core/src/modules/compensation/job-offers.ts
